+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * Python >=3.9 required
 * `isinstance` and generic dispatch optimized
 
+### Added
+* Custom instance checks
+* `subtype` and `parametric` utilities
+
 ## [1.10](https://pypi.org/project/multimethod/1.10/) - 2023-09-21
 ### Changed
 * Python >=3.8 required